requirements.txt
Flask==3.1.0
pywebview
pyinstaller
